refactor: implement exception handling for Department endpoints

+++ src/main/java/com/demo/spring/demo_spring/services/DepartmentService.java
@@ -2,8 +2,12 @@ package com.demo.spring.demo_spring.services;
 
 import com.demo.spring.demo_spring.dto.DepartmentDTO;
 import com.demo.spring.demo_spring.entities.Department;
+import com.demo.spring.demo_spring.exceptions.DatabaseException;
+import com.demo.spring.demo_spring.exceptions.ResourceNotFoundException;
 import com.demo.spring.demo_spring.repositories.DepartmentRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,7 +22,7 @@ public class DepartmentService {
     @Transactional(readOnly = true)
     public DepartmentDTO findById(Long id) {
         Department department = repository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Department not found"));
+                .orElseThrow(() -> new ResourceNotFoundException("Department not found"));
         return new DepartmentDTO(department);
     }
 
@@ -38,14 +42,29 @@ public class DepartmentService {
 
     @Transactional
     public void delete(Long id) {
-        repository.deleteById(id);
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("ID dont exists");
+        }
+        try {
+            repository.deleteById(id);
+
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Integrity violation");
+        }
     }
 
     @Transactional
     public DepartmentDTO update(Long id, DepartmentDTO dto) {
-        Department entity = repository.getReferenceById(id);
-        entity.setName(dto.getName());
-        entity = repository.save(entity);
-        return new DepartmentDTO(entity);
+
+        try {
+            Department entity = repository.getReferenceById(id);
+            entity.setName(dto.getName());
+            entity = repository.save(entity);
+            return new DepartmentDTO(entity);
+
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("ID dont exists");
+        }
     }
 }
